崩溃问题解决

+++ src/com/suken/bridgedetection/util/AlarmReceiver.java
@@ -1,5 +1,6 @@
 package com.suken.bridgedetection.util;
 
+import com.suken.bridgedetection.BridgeDetectionApplication;
 import com.suken.bridgedetection.storage.CheckFormAndDetailDao;
 import com.suken.bridgedetection.storage.SdxcFormAndDetailDao;
 
@@ -12,9 +13,11 @@ public class AlarmReceiver extends BroadcastReceiver {
 	@Override
 	public void onReceive(Context context, Intent intent) {
 		if (intent.getAction().equals("android.suken.action.checkdata")) {
-			// 清理数据
-			new CheckFormAndDetailDao().deleteAllLocalData();
-			new SdxcFormAndDetailDao().deleteAllLocalData();
+			if(BridgeDetectionApplication.mCurrentUser != null) {
+				// 清理数据
+				new CheckFormAndDetailDao().deleteAllLocalData();
+				new SdxcFormAndDetailDao().deleteAllLocalData();
+			}
 		} else if (intent.getAction().equals(Intent.ACTION_BOOT_COMPLETED)) {
 			// 开机自启动则启动
 			UiUtil.setAlarm(context);
